requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-ollama==0.1.0
langchain-community==0.1.0
streamlit==1.29.0
python-dotenv==1.0.0
requests==2.31.0